==> code/01_process_gene_x_sample.ipynb <==
%load_ext autoreload
%autoreload 2

from __init__ import *

feature_x_sample = pd.read_csv(
    os.path.join("../input/", "feature_x_sample.tsv"), sep="\t", index_col=0
)

feature_x_sample

kraft.plot.plot_plotly(
    {
        "layout": {"xaxis": {"title": "Value"}},
        "data": [
            {"type": "histogram", "name": sample, "x": values}
            for sample, values in feature_x_sample.items()
        ],
    }
)
