==> code/1.ipynb <==
%load_ext autoreload
%autoreload 2

from __init__ import *

nu_fe_sa = pd.read_csv("../input/number_feature_sample.tsv", "\t", index_col=0)

nu_fe_sa.columns.name = "Sample"

if 0 < len(SETTING["pr"]):

    nu_fe_sa = kraft.feature_by_sample.process(nu_fe_sa, **SETTING["pr"])

nu_fe_sa.to_csv("../output/number_feature_sample.tsv", "\t")

kraft.plot.plot_histogram(
    [se for la, se in nu_fe_sa.iteritems()],
)

nu_fe_sa.loc[SETTING["fe_"], :]
