==> code/1_all.ipynb <==
%load_ext autoreload
%autoreload 2

from __init__ import *

n_ = sorted(
    (n for n in os.listdir() if re.match(r"^[^._].+\.ipynb$", n)),
    key=lambda n: int(n.split("_")[0]),
)

n_.remove("1_all.ipynb")

for n in n_:

    kraft.shell.command(
        "jupyter nbconvert --execute --ExecutePreprocessor.timeout=-1 --inplace {}".format(
            n
        )
    )
