==> code/2.ipynb <==
%load_ext autoreload
%autoreload 2

from __init__ import *

nu_fe_sa = pd.read_csv("../output/number_feature_sample.tsv", "\t", index_col=0)

bi_fe_sa = pd.read_csv("../input/01_feature_sample.tsv", "\t", index_col=0)

bi_co_sa = pd.DataFrame(columns=nu_fe_sa.columns)

for co in SETTING["co_"]:

    if isinstance(co, str):

        bi_co_sa.loc[co, :] = bi_fe_sa.loc[co, :]

    else:

        fe0_, fe1_ = co

        co = "{} vs {}".format(",".join(fe0_), ",".join(fe1_))

        for fe_, it in [[fe0_, 0], [fe1_, 1]]:

            bi_co_sa.loc[co, bi_fe_sa.loc[fe_].all()] = it

bi_co_sa.index.name = "Comparison"

bi_co_sa.to_csv("../output/01_comparison_sample.tsv", "\t")

bi_co_sa
