==> code/2_feature_x_sample.ipynb <==
%load_ext autoreload
%autoreload 2

from __init__ import *

fxs = pd.read_csv("../input/feature_x_sample.tsv", sep="\t", index_col=0)

fxs

if 0 < len(S["process"]):

    fxs = kraft.pd.process(fxs, **S["process"])

kraft.plot.plot_histogram(
    [v for s, v in fxs.iteritems()],
    [fxs.index] * fxs.shape[1],
    fxs.columns,
)

fxs.loc[S["peek_feature_"], :]
