==> code/2_process_gene_x_sample.ipynb <==
%load_ext autoreload
%autoreload 2

from environment import *

with open("setting.json") as io:

    SETTING = json.load(io)

PATH = make_path_dict(SETTING)

gene_x_sample = pd.read_csv(PATH["gene_x_sample.tsv"], sep="\t", index_col=0)

gene_x_sample__processed = kraft.process_feature_x_sample(
    gene_x_sample,
    features_to_drop=SETTING["features_to_drop"],
    samples_to_drop=SETTING["samples_to_drop"],
    nanize=SETTING["nanize"],
    drop_axis=SETTING["drop_axis"],
    max_na=SETTING["max_na"],
    min_n_not_na_unique_value=SETTING["min_n_not_na_unique_value"],
    shift_as_necessary_to_achieve_min_before_logging=SETTING[
        "shift_as_necessary_to_achieve_min_before_logging"
    ],
    log_base=SETTING["log_base"],
    normalization_axis=SETTING["normalization_axis"],
    normalization_method=SETTING["normalization_method"],
    clip_min=SETTING["clip_min"],
    clip_max=SETTING["clip_max"],
    plot_heat_map_max_size=SETTING["plot_heat_map_max_size"],
    plot_histogram_max_size=SETTING["plot_histogram_max_size"],
    plot_rug_max_size=SETTING["plot_rug_max_size"],
)

gene_x_sample__processed.to_csv(PATH["gene_x_sample.processed.tsv"], sep="\t")

gene_x_sample__processed

pl.offline.iplot(
    {
        "layout":{
            title={"text":"Distribution of Sample Values"}, "xaxis":{"title":"Value"}
        },
        "data":[
            {"type":="histogram", "name":column_name, "x":column_values}
            for column_name, column_values in gene_x_sample__processed.items()
        ],
    }
)
