==> code/3_comparison_x_sample.ipynb <==
%load_ext autoreload
%autoreload 2

from __init__ import *

fxs = pd.read_csv("../input/feature_x_sample.tsv", sep="\t", index_col=0)

bfxs = pd.read_csv("../input/binary_feature_x_sample.tsv", sep="\t", index_col=0)

cxs = pd.DataFrame(columns=fxs.columns)

for c in S["comparison_"]:

    if isinstance(c, str):

        cxs.loc[c, :] = bfxs.loc[c, :]

    else:

        g0, g1 = c

        c = "{} vs {}".format(",".join(g0), ",".join(g1))

        for f_, i in [[g0, 0], [g1, 1]]:

            cxs.loc[c, bfxs.loc[f_].all()] = i

cxs.index.name = "Comparison"

cxs.to_csv("../output/comparison_x_sample.tsv", sep="\t")

cxs
