==> code/3_find_gene_set.ipynb <==
%load_ext autoreload
%autoreload 2

from environment import *

with open("setting.yaml") as yaml_file:

    SETTING = yaml.load(yaml_file)

PATH = make_path_dict(SETTING)

gene_sets = ccal.read_gmts(SETTING["gene_set_file_paths"])

select = np.full(gene_sets.index.size, False)

select |= gene_sets.index.str.lower().str.contains("metaboli")

select |= gene_sets.index.str.lower().str.contains("pyrimidine")

gene_sets__selected = gene_sets.index[select]

print(gene_sets__selected.size)

sorted(gene_sets__selected)

ccal.write_gmt(gene_sets.loc[gene_sets__selected], "/home/kwatme/metabolism.gmt")
