==> code/04.ipynb <==
%load_ext autoreload
%autoreload 2

from __init__ import *

feature_x_sample = pd.read_csv("../input/feature_x_sample.tsv", sep="\t", index_col=0)

comparison_x_sample = pd.read_csv(
    "../output/comparison_x_sample.tsv", sep="\t", index_col=0
)

sets = kraft.gmt.read_many(kraft.path.list_("../input/set/"))

set_x_information = pd.DataFrame(
    data={
        "Selected": False,
        "N": tuple(len(features) for features in sets.values()),
        "n": tuple(
            sum(feature in feature_x_sample.index for feature in features)
            for features in sets.values()
        ),
    },
    index=sets,
)

set_x_information.loc[:, "."] = (
    set_x_information.loc[:, "n"].to_numpy() / set_x_information.loc[:, "N"].to_numpy()
)

set_x_information

len(
    set(
        kraft.series.get_extreme_labels(
            set_x_information.loc[:, "n"],
            ">",
            low_and_high=(None, SETTING["set_n"]),
            plot=False,
        )
    ).intersection(
        set(
            kraft.series.get_extreme_labels(
                set_x_information.loc[:, "."],
                ">",
                low_and_high=(None, SETTING["set_."]),
                plot=False,
            )
        )
    )
)

set_x_information.loc[
    set(
        kraft.series.get_extreme_labels(
            set_x_information.loc[:, "n"],
            ">",
            low_and_high=(None, SETTING["set_n"]),
            plot=False,
        )
    )
    & set(
        kraft.series.get_extreme_labels(
            set_x_information.loc[:, "."],
            ">",
            low_and_high=(None, SETTING["set_."]),
            plot=False,
        )
    ),
    "Selected",
] = True

feature_pattern = "YAP[0-9]*"

boundary_pattern = "[\s_]"

for set_ in sets:

    if re.match(
        r"^{0}$|.*{1}{0}{1}.*|^{0}{1}.*|.*{1}{0}$".format(
            feature_pattern, boundary_pattern
        ),
        set_,
    ):

        print(set_)

        set_x_information.loc[set_, "Selected"] = True

set_x_information.loc[SETTING["sets_to_peek"], "Selected"] = True

set_x_information.loc[:, "Selected"].sum()

set_x_information.to_csv("../output/set_x_information.tsv", sep="\t")

set_x_information

set_x_information.reindex(SETTING["sets_to_peek"])
