==> code/4.ipynb <==
%load_ext autoreload
%autoreload 2

from __init__ import *

nu_fe_sa = pd.read_csv("../output/number_feature_sample.tsv", "\t", index_col=0)

bi_co_sa = pd.read_csv("../output/01_comparison_sample.tsv", "\t", index_col=0)

se_fe_ = kraft.gmt.read(kraft.path.list("../input/set/"))

for se in se_fe_:

    if re.match(
        r"^{0}$|.*{1}{0}{1}.*|^{0}{1}.*|.*{1}{0}$".format("FATIGUE[0-9]*", "[\s_]"),
        se,
    ):

        print(se)

fe__ = se_fe_.values()

st_se_in = pd.DataFrame(
    {
        "Selected": False,
        "N": [len(fe_) for fe_ in fe__],
        "n": [sum(fe in nu_fe_sa.index for fe in fe_) for fe_ in fe__],
    },
    se_fe_,
)

st_se_in.loc[:, "."] = st_se_in.loc[:, "n"] / st_se_in.loc[:, "N"]

st_se_in

st_se_in.loc[
    kraft.array.check_is_extreme(
        st_se_in.loc[:, "n"].values,
        ">",
        th_=[None, SETTING["n_ge"]],
    )
    & kraft.array.check_is_extreme(
        st_se_in.loc[:, "."].values,
        ">",
        th_=[None, SETTING["._ge"]],
    ),
    "Selected",
] = True

assert (1 / 2 < st_se_in.loc[SETTING["se_"], "."]).all(), st_se_in.loc[
    SETTING["se_"], "."
]

st_se_in.loc[SETTING["se_"], "Selected"] = True

st_se_in.to_csv("../output/statistic_set_information.tsv", "\t")

print(st_se_in.loc[:, "Selected"].sum())

st_se_in

st_se_in.loc[SETTING["se_"], :].sort_values(".")
