==> code/4_compare_feature.ipynb <==
%load_ext autoreload
%autoreload 2

from __init__ import *

fxs = pd.read_csv("../input/feature_x_sample.tsv", sep="\t", index_col=0)

cxs = pd.read_csv("../output/comparison_x_sample.tsv", sep="\t", index_col=0)

def get_log_ratio(is_, v):

    v = kraft.array.log(kraft.array.shift_min(v, 1))

    return v[is_ == 1].mean() - v[is_ == 0].mean()

kwarg_ = {
    "v_data_type": "binary",
    "plot_std": S["plot_std"],
}

for c, is_ in cxs.iterrows():

    is_.dropna(inplace=True)

    m = np.unique(is_, return_counts=True)[1].min()

    for f, s in (
        (get_log_ratio, 1),
        (kraft.information.get_ic, 2),
    ):

        if m < s:

            continue

        d = "../output/compare_feature/{}/{}/".format(c, f.__name__)

        kraft.path.make(d)

        st = kraft.function_heat_map.make(
            is_,
            fxs,
            f,
            n_job=S["n_job"],
            n_sample=S["n_sample"],
            n_shuffle=S["n_shuffle"],
            n_plot=S["n_extreme"],
            title="All ({})".format(f.__name__),
            directory_path=d,
            **kwarg_,
        )

        kraft.function_heat_map.make(
            is_,
            fxs.reindex(index=S["peek_feature_"]),
            st,
            n_plot=None,
            title="Peek ({})".format(f.__name__),
            **kwarg_,
        )
