==> code/05_gsea.ipynb <==
%load_ext autoreload
%autoreload 2

from __init__ import *

feature_x_sample = pd.read_csv("../input/feature_x_sample.tsv", sep="\t", index_col=0)

comparison_x_sample = pd.read_csv(
    "../output/comparison_x_sample.tsv", sep="\t", index_col=0
)

sets = kraft.gmt.read_many(kraft.path.list_("../input/feature_set/"))

set_x_information = pd.read_csv(
    "../output/set_x_information.tsv", sep="\t", index_col=0
)

for comparison in comparison_x_sample.index.to_numpy():

    directory_path = os.path.join("../output/sea", kraft.path.clean(comparison), "")

    kraft.path.path(directory_path)

    scores = pd.read_csv(
        os.path.join(
            "../output/find_differentially_expressed_feature",
            kraft.path.clean(comparison),
            "all_{}".format(SETTING["sea_function"]),
            "scores.tsv",
        ),
        sep="\t",
        index_col=0,
    )

    score_ = scores.loc[:, "Score"]

    score_.name = SETTING["sea_function"]

    set_scores = kraft.sea.score_samples_and_sets(
        score_.to_frame(),
        {
            set_: sets[set_]
            for set_ in set_x_information.index.to_numpy()[
                set_x_information.loc[:, "Selected"].to_numpy()
            ]
        },
        n_job=SETTING["n_job"],
    ).squeeze()

    set_scores.to_csv(
        os.path.join(directory_path, "{}.tsv".format(SETTING["sea_function"])),
        sep="\t",
    )

"../output/sea/"

set_score.name = "GSEA Score ({})".format(SETTING["gsea_statistic"])


    print(set_x_information.loc[set_score.index])

    kraft.plot_point_and_annotate(
        None,
        set_score,
        "y",
        annotation=(("Peek", SETTING["sets_to_peek"], 8, "#20d9ba"),),
        title={"text": target_name},
        html_file_path=os.path.join(
            directory_path,
            "{}.plot_point_and_annotate.html".format(
                SETTING["gsea_match_function_name"]
            ),
        ),
    )

    directory_path = os.path.join(
        directory_path, "{}.mountain".format(SETTING["gsea_match_function_name"])
    )

    kraft.establish_path(directory_path, "directory")

    for set_name in SETTING["sets_to_peek"]:

        kraft.single_sample_gsea(
            feature_score,
            sets.loc[set_name],
            statistic=SETTING["gsea_statistic"],
            title="{}<br>{}".format(target_name, set_name),
            html_file_path=os.path.join(
                directory_path, "{}.html".format(set_name)
            ),
        )
