==> code/5_set_x_information.ipynb <==
%load_ext autoreload
%autoreload 2

from __init__ import *

fxs = pd.read_csv("../input/feature_x_sample.tsv", sep="\t", index_col=0)

cxs = pd.read_csv("../output/comparison_x_sample.tsv", sep="\t", index_col=0)

stf_ = kraft.gmt.read(kraft.path.list_absolute("../input/set/"))

f__ = stf_.values()

sxi = pd.DataFrame(
    {
        "Selected": False,
        "N": [len(f_) for f_ in f__],
        "n": [sum(f in fxs.index for f in f_) for f_ in f__],
    },
    index=stf_,
)

sxi.loc[:, "."] = sxi.loc[:, "n"] / sxi.loc[:, "N"]

sxi

sxi.loc[
    kraft.array.check_is_extreme(
        sxi.loc[:, "n"].to_numpy(),
        ">",
        low_and_high=[None, S["n_set"]],
    )
    & kraft.array.check_is_extreme(
        sxi.loc[:, "."].to_numpy(),
        ">",
        low_and_high=[None, S["._set"]],
    ),
    "Selected",
] = True

sxi.loc[S["peek_set_"], "Selected"] = True

sxi.loc[:, "Selected"].sum()

sxi.to_csv("../output/set_x_information.tsv", sep="\t")

sxi

sxi.loc[S["peek_set_"], :]

for s in stf_:

    if re.match(
        r"^{0}$|.*{1}{0}{1}.*|^{0}{1}.*|.*{1}{0}$".format("FATIGUE[0-9]*", "[\s_]"),
        s,
    ):

        print(s)
