==> code/06_make_gene_set_x_sample.ipynb <==
%load_ext autoreload
%autoreload 2

from environment import *

with open("setting.json") as io:

    SETTING = json.load(io)

PATH = make_path_dict(SETTING)

gene_x_sample = pd.read_csv(PATH["gene_x_sample.processed.tsv"], sep="\t", index_col=0)

gene_sets = kraft.read_gmts(SETTING["gene_set_file_paths"])

gene_set_x_information = pd.read_csv(
    PATH["gene_set_x_information.tsv"], sep="\t", index_col=0
)

gene_set_x_sample = kraft.single_sample_gseas(
    gene_x_sample,
    gene_sets.loc[
        gene_set_x_information.index[gene_set_x_information["Good"].astype(bool)]
    ],
    statistic=SETTING["gsea_statistic"],
    n_job=SETTING["n_job"],
    file_path=PATH["gene_set_x_sample.tsv"],
)

kraft.summarize_feature_x_sample(gene_set_x_sample)

gene_set_x_sample
