==> code/6_score_set.ipynb <==
%load_ext autoreload
%autoreload 2

from __init__ import *

cxs = pd.read_csv("../output/comparison_x_sample.tsv", sep="\t", index_col=0)

stf_ = kraft.gmt.read(kraft.path.list_absolute("../input/set/"))

sxi = pd.read_csv("../output/set_x_information.tsv", sep="\t", index_col=0)

for c in cxs.index:

    fs_ = pd.read_csv(
        "../output/compare_feature/{}/{}/statistic.tsv".format(
            c, S["score_for_scoring_set"]
        ),
        sep="\t",
        index_col=0,
    ).loc[:, "Score"]

    d = "../output/score_set/{}/".format(c)

    kraft.path.make(d)

    ss_ = kraft.score_set.score_samples_and_sets(
        fs_.to_frame(),
        {s: stf_[s] for s in sxi.index[sxi.loc[:, "Selected"]]},
    ).squeeze()

    ss_.to_csv("{}/set_score_.tsv".format(d), sep="\t")

    for s in S["peek_set_"]:

        kraft.score_set.score_sample_and_set(
            fs_,
            stf_[s],
            title="{}<br>{}".format(c, s),
            file_path="{}/{}.html".format(d, s),
        )
