==> code/08.ipynb <==
%load_ext autoreload
%autoreload 2

from __init__ import *

feature_x_sample = pd.read_csv("../input/feature_x_sample.tsv", sep="\t", index_col=0)

comparison_x_sample = pd.read_csv(
    "../output/comparison_x_sample.tsv", sep="\t", index_col=0
)

sets = kraft.gmt.read_many(kraft.path.list_("../input/set/"))

set_x_sample = pd.read_csv("../output/set_x_sample.tsv", sep="\t", index_col=0)

for comparison, groups in comparison_x_sample.iterrows():

    groups.dropna(inplace=True)

    min_group_size = np.unique(groups.to_numpy(), return_counts=True)[1].min()

    function, group_size = "get_ic", 2

    if min_group_size < group_size:

        continue

    path_template = "../output/compare_{{}}/{}/{}/scores.tsv".format(
        comparison, function
    )

    scores = pd.concat(
        (
            pd.read_csv(path_template.format("feature"), sep="\t", index_col=0,),
            pd.read_csv(path_template.format("set"), sep="\t", index_col=0,),
        )
    )

    directory_path = "../output/expand_set/{}/".format(comparison)

    kraft.path.path(directory_path)

    for set_ in SETTING["sets_to_peek"]:

        kraft.function_heat_map.make(
            groups,
            pd.concat(
                (
                    feature_x_sample.reindex(labels=sets[set_]),
                    set_x_sample.reindex(labels=(set_,)),
                )
            ),
            None,
            scores=scores,
            n_extreme=None,
            se_data_type="binary",
            plot_std=SETTING["plot_std"],
            title=set_,
            directory_path=directory_path,
        )
