==> code/9.ipynb <==
%load_ext autoreload
%autoreload 2

from __init__ import *

bi_co_sa = pd.read_csv("../output/01_comparison_sample.tsv", "\t", index_col=0)

nu_fe_sa = pd.read_csv("../output/number_feature_sample.tsv", "\t", index_col=0)

sc_se_sa = pd.read_csv("../output/score_set_sample.tsv", "\t", index_col=0)

title = "Summary"

for (co, ta), bl_se_ in zip(bi_co_sa.iterrows(), SETTING["su_"]):

    ta.dropna(inplace=True)

    fu = kraft.information.get_ic

    pa = "../output/summarize_set/{}/".format(kraft.path.clean(co))

    kraft.path.make(pa)

    bu_ = []

    for bl, se_ in bl_se_.items():

        bu_.append(
            {
                "name": bl,
                "data": pd.concat([nu_fe_sa, sc_se_sa]).reindex(se_),
                "type": "continuous",
                "statistic": read_statistic(kraft.path.clean(co), fu.__name__),
            }
        )

    kraft.function_heat_map.summarize(
        ta,
        bu_,
        ty="binary",
        st=SETTING["st"],
        title=title,
        pa="{}{}.html".format(pa, kraft.path.clean(title)),
    )
