==> code/9_expand_set.ipynb <==
%load_ext autoreload
%autoreload 2

from __init__ import *

fxs = pd.read_csv("../input/feature_x_sample.tsv", sep="\t", index_col=0)

cxs = pd.read_csv("../output/comparison_x_sample.tsv", sep="\t", index_col=0)

stf_ = kraft.gmt.read(kraft.path.list_absolute("../input/set/"))

sxs = pd.read_csv("../output/set_x_sample.tsv", sep="\t", index_col=0)

for c, is_ in cxs.iterrows():

    is_.dropna(inplace=True)

    m = np.unique(is_, return_counts=True)[1].min()

    f, s = kraft.information.get_ic, 2

    if m < s:

        continue

    t = "../output/compare_{{}}/{}/{}/statistic.tsv".format(c, f.__name__)

    s_ = pd.concat(
        (
            pd.read_csv(
                t.format("feature"),
                sep="\t",
                index_col=0,
            ),
            pd.read_csv(
                t.format("set"),
                sep="\t",
                index_col=0,
            ),
        )
    )

    d = "../output/expand_set/{}/".format(c)

    kraft.path.make(d)

    for s in S["peek_set_"]:

        kraft.function_heat_map.make(
            is_,
            pd.concat(
                [
                    fxs.loc[fxs.index.intersection(stf_[s])],
                    sxs.loc[[s]],
                ]
            ),
            s_,
            n_plot=None,
            v_data_type="binary",
            plot_std=S["plot_std"],
            title=s,
            directory_path=d,
        )
